fix: move nlj operator into physical subfolder
