+++ core/src/main/java/io/substrait/relation/AbstractRelVisitor.java
@@ -1,6 +1,7 @@
 package io.substrait.relation;
 
 import io.substrait.relation.physical.HashJoin;
+import io.substrait.relation.physical.NestedLoopJoin;
 
 public abstract class AbstractRelVisitor<OUTPUT, EXCEPTION extends Exception>
     implements RelVisitor<OUTPUT, EXCEPTION> {
