+++ core/src/main/java/io/substrait/relation/RelCopyOnWriteVisitor.java
@@ -8,6 +8,8 @@ import io.substrait.expression.FunctionArg;
 import io.substrait.expression.ImmutableFieldReference;
 import io.substrait.relation.physical.HashJoin;
 import io.substrait.relation.physical.ImmutableHashJoin;
+import io.substrait.relation.physical.ImmutableNestedLoopJoin;
+import io.substrait.relation.physical.NestedLoopJoin;
 import io.substrait.type.Type;
 import java.util.ArrayList;
 import java.util.List;
