+++ core/src/test/java/io/substrait/type/proto/ExtensionRoundtripTest.java
@@ -16,7 +16,6 @@ import io.substrait.relation.Filter;
 import io.substrait.relation.Join;
 import io.substrait.relation.LocalFiles;
 import io.substrait.relation.NamedScan;
-import io.substrait.relation.NestedLoopJoin;
 import io.substrait.relation.Project;
 import io.substrait.relation.ProtoRelConverter;
 import io.substrait.relation.Rel;
@@ -24,6 +23,7 @@ import io.substrait.relation.Set;
 import io.substrait.relation.Sort;
 import io.substrait.relation.VirtualTableScan;
 import io.substrait.relation.physical.HashJoin;
+import io.substrait.relation.physical.NestedLoopJoin;
 import io.substrait.relation.utils.StringHolder;
 import io.substrait.relation.utils.StringHolderHandlingProtoRelConverter;
 import io.substrait.type.NamedStruct;
